middleware/requirements.txt
fastapi
uvicorn
httpx
aiohttp
opensearch-dsl
pytest
python-dotenv
watchdog
PyMuPDF
beautifulsoup4
xmltodict
pytest-asyncio
ollama
pandas
python-multipart